build(probabilistic_occupancy_grid_map): move header files to the probabilistic_occupancy_grid_map folder to avoid conflicts
