+++ perception/probabilistic_occupancy_grid_map/include/probabilistic_occupancy_grid_map/fusion/synchronized_grid_map_fusion_node.hpp
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
-#define FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
+#ifndef PROBABILISTIC_OCCUPANCY_GRID_MAP__FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
+#define PROBABILISTIC_OCCUPANCY_GRID_MAP__FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
 
-#include "fusion/single_frame_fusion_policy.hpp"
-#include "pointcloud_based_occupancy_grid_map/occupancy_grid_map_fixed.hpp"
-#include "updater/occupancy_grid_map_log_odds_bayes_filter_updater.hpp"
-#include "updater/occupancy_grid_map_updater_interface.hpp"
+#include "probabilistic_occupancy_grid_map/fusion/single_frame_fusion_policy.hpp"
+#include "probabilistic_occupancy_grid_map/updater/occupancy_grid_map_log_odds_bayes_filter_updater.hpp"
+#include "probabilistic_occupancy_grid_map/updater/occupancy_grid_map_updater_interface.hpp"
+
+#include "probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/occupancy_grid_map_fixed.hpp"
 
 #include <builtin_interfaces/msg/time.hpp>
 #include <rclcpp/rclcpp.hpp>
@@ -124,4 +125,4 @@ private:
 
 }  // namespace synchronized_grid_map_fusion
 
-#endif  // FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
+#endif  // PROBABILISTIC_OCCUPANCY_GRID_MAP__FUSION__SYNCHRONIZED_GRID_MAP_FUSION_NODE_HPP_
