+++ perception/probabilistic_occupancy_grid_map/include/probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/pointcloud_based_occupancy_grid_map_node.hpp
@@ -12,16 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef LASERSCAN_BASED_OCCUPANCY_GRID_MAP__LASERSCAN_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
-#define LASERSCAN_BASED_OCCUPANCY_GRID_MAP__LASERSCAN_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
+#ifndef PROBABILISTIC_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
+#define PROBABILISTIC_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
 
-#include "laserscan_based_occupancy_grid_map/occupancy_grid_map.hpp"
-#include "updater/occupancy_grid_map_binary_bayes_filter_updater.hpp"
-#include "updater/occupancy_grid_map_updater_interface.hpp"
+#include "probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/occupancy_grid_map_base.hpp"
+#include "probabilistic_occupancy_grid_map/updater/occupancy_grid_map_binary_bayes_filter_updater.hpp"
+#include "probabilistic_occupancy_grid_map/updater/occupancy_grid_map_updater_interface.hpp"
 
 #include <builtin_interfaces/msg/time.hpp>
 #include <laser_geometry/laser_geometry.hpp>
 #include <rclcpp/rclcpp.hpp>
+#include <tier4_autoware_utils/ros/debug_publisher.hpp>
+#include <tier4_autoware_utils/system/stop_watch.hpp>
 
 #include <sensor_msgs/msg/laser_scan.hpp>
 #include <sensor_msgs/point_cloud2_iterator.hpp>
@@ -39,6 +41,7 @@
 namespace occupancy_grid_map
 {
 using builtin_interfaces::msg::Time;
+using costmap_2d::OccupancyGridMapInterface;
 using costmap_2d::OccupancyGridMapUpdaterInterface;
 using laser_geometry::LaserProjection;
 using nav2_costmap_2d::Costmap2D;
@@ -48,46 +51,35 @@ using sensor_msgs::msg::PointCloud2;
 using tf2_ros::Buffer;
 using tf2_ros::TransformListener;
 
-class LaserscanBasedOccupancyGridMapNode : public rclcpp::Node
+class PointcloudBasedOccupancyGridMapNode : public rclcpp::Node
 {
 public:
-  explicit LaserscanBasedOccupancyGridMapNode(const rclcpp::NodeOptions & node_options);
+  explicit PointcloudBasedOccupancyGridMapNode(const rclcpp::NodeOptions & node_options);
 
 private:
-  PointCloud2::SharedPtr convertLaserscanToPointCLoud2(const LaserScan::ConstSharedPtr & input);
-  void onLaserscanPointCloud2WithObstacleAndRaw(
-    const LaserScan::ConstSharedPtr & input_laserscan_msg,
+  void onPointcloudWithObstacleAndRaw(
     const PointCloud2::ConstSharedPtr & input_obstacle_msg,
     const PointCloud2::ConstSharedPtr & input_raw_msg);
   OccupancyGrid::UniquePtr OccupancyGridMapToMsgPtr(
     const std::string & frame_id, const Time & stamp, const float & robot_pose_z,
     const Costmap2D & occupancy_grid_map);
-  inline void onDummyPointCloud2(const LaserScan::ConstSharedPtr & input)
-  {
-    PointCloud2 dummy;
-    sensor_msgs::PointCloud2Modifier modifier(dummy);
-    modifier.setPointCloud2FieldsByString(1, "xyz");
-    dummy.header = input->header;
-    passthrough_.add(std::make_shared<PointCloud2>(dummy));
-  }
 
 private:
   rclcpp::Publisher<OccupancyGrid>::SharedPtr occupancy_grid_map_pub_;
-  message_filters::Subscriber<LaserScan> laserscan_sub_;
   message_filters::Subscriber<PointCloud2> obstacle_pointcloud_sub_;
   message_filters::Subscriber<PointCloud2> raw_pointcloud_sub_;
-  message_filters::PassThrough<PointCloud2> passthrough_;
+  std::unique_ptr<tier4_autoware_utils::StopWatch<std::chrono::milliseconds>> stop_watch_ptr_{};
+  std::unique_ptr<tier4_autoware_utils::DebugPublisher> debug_publisher_ptr_{};
 
   std::shared_ptr<Buffer> tf2_{std::make_shared<Buffer>(get_clock())};
   std::shared_ptr<TransformListener> tf2_listener_{std::make_shared<TransformListener>(*tf2_)};
 
-  using SyncPolicy = message_filters::sync_policies::ExactTime<LaserScan, PointCloud2, PointCloud2>;
+  using SyncPolicy = message_filters::sync_policies::ExactTime<PointCloud2, PointCloud2>;
   using Sync = message_filters::Synchronizer<SyncPolicy>;
   std::shared_ptr<Sync> sync_ptr_;
 
-  LaserProjection laserscan2pointcloud_converter_;
-
-  std::shared_ptr<OccupancyGridMapUpdaterInterface> occupancy_grid_map_updater_ptr_;
+  std::unique_ptr<OccupancyGridMapInterface> occupancy_grid_map_ptr_;
+  std::unique_ptr<OccupancyGridMapUpdaterInterface> occupancy_grid_map_updater_ptr_;
 
   // ROS Parameters
   std::string map_frame_;
@@ -98,8 +90,9 @@ private:
   double min_height_;
   double max_height_;
   bool enable_single_frame_mode_;
+  bool filter_obstacle_pointcloud_by_raw_pointcloud_;
 };
 
 }  // namespace occupancy_grid_map
 
-#endif  // LASERSCAN_BASED_OCCUPANCY_GRID_MAP__LASERSCAN_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
+#endif  // PROBABILISTIC_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP__POINTCLOUD_BASED_OCCUPANCY_GRID_MAP_NODE_HPP_
